backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
redis==5.0.1
celery==5.3.4
google-generativeai==0.3.2
# Azure ML (optional - only install if using Azure)
# azure-ai-ml==1.11.1
# azure-identity==1.15.0

# Local ML alternatives (commented out for Windows compatibility)
# scikit-learn==1.3.2
# numpy==1.24.3
# pandas==2.1.4
# joblib==1.3.2
requests==2.31.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2